refactor(l5): move rotating-file from src/ into server/logging (issue 864)
One move. `src/rotating-file.ts` was a 246-line implementation sitting in
src/ while every other logging sink already lived under server/logging/.
The module and its test move to server/logging/, and `src/rotating-file.ts`
becomes an L5 re-export shim so existing importers keep resolving unchanged;
the shim retires with src/ at L6.

The test file's single oversized describe is split into two siblings,
`createRotatingFileSink` and `createRotatingFileSink edge sizes`, so each
stays under the per-function line rule the staged-file lint enforces. It is
not a .pg.test.ts and has no entry in scripts/assert-db-tests-ran.ts, so no
manifest change is needed.

One doc-comment path in server/tools/dependency-arrival.test.ts is repointed
because it named `src/rotating-file.test.ts`, which no longer exists after
this move. The comparable reference in server/observability/trace-sink-health.ts
still names `src/rotating-file.ts`, which remains accurate while the shim is
there, so that file is untouched.

That repointed file also carries pre-existing repo-wide prettier drift
(issue 753): the committed formatting predates the installed prettier, so
the pre-commit gate refuses the file whole rather than by changed line. The
gate is satisfied by formatting the staged file, which is what the other L5
lanes did, so the diff there is one comment path plus reflowed formatting
with no behavior change.

Deferred, with reasons:

- with-logging: the RETARGET candidate is server/logging/decorate.ts, which
  is already under server/. Moving it is a different change with different
  importers, not part of this move.
- audit-log: it reads four OPENBRAIN_MCP_AUDIT_* keys that are absent from
  server/config.ts. Under server/ the no-process.env rule applies, so the
  move requires threading those keys through config first. That is its own
  lane rather than a silent config change smuggled into a file move.

Verified:
- bunx tsc --noEmit -> exit 0
- oxlint --deny-warnings on the four staged paths -> exit 0
- bash scripts/done-means/864-moved-out-of-src.sh -> judged=6 shims, PASS
- bun run test:isolated on rotating-file, dependency-arrival, main.pg
  -> 24 pass, 0 fail
